Makefile: Fixes integration test coverage reporting
When passing build tags to the go test tool, it was believed that only tests with those tags would run.
However including build tags executes all tests and not just that of the passed tag.

This changes isolates the integration test and generates coverage for that only, restoring accuracy
of the coverage we gain by using the integration test framework.