adapter: Refactor - Moves all non-recoverable errors to handler function
This change moves any code that is required as a pre-condition to authorization into the top level function.
It allows us to exit early in case of any error that is outside our control before parsing the request
and/or contacting 3scale backend to determine if the auth request should be granted.