docs: Document 1.1 dependency and additional auth pattern config
This change alerts the operator to the fact that the config has no changed to use features
from Istio version 1.1 which becomes a prerequisite.

Also included are updates to the README to describe how to configure the various authentication patterns
and how to adjust the config to enforce various authn methods.