testdata, templating, config: This change updates the name of the authorization template from authorization to threescale-authorization
This change was made to match the naming convention in the Maistra project to avoid confusion when a user goes from installing
the adapter as a standalone resource, to installing and configuring the adapter via Maistra.

Updates the integration testdata and makes the relevant changes in the templating packages and the example templates.