docs: Updates for istio 1.1.0 changes and some fixes to existing docs.
As of Istio 1.1.0, policies have been disabled with the default install. These need to
be enabled for the adapter to work correctly so we link off to upstream in the README

The patch command for the deployemtnwas misleading, it used SVC_ID as the name when in reality
it needed the unique ID generated by the cli. Updated to avoid confusion.

Updates to the sample custom resources to match the names in the templates as per change from
authorization template to threescale-authorization.