server,cli: Support passing the binary version via build flags
It can be useful to see what version of the code has been compiled as is running.
This change allows the server and config generator applications to be passed a 'version'
variable via a build flag.

Change to cli allows outputting of version via flag.
gRPC sever now logs the version on startup.