test: adapter: oidc Adds test for provider creation
Change adds a unit test for providing an 'OIDC Handler' with a given url which will be used via
the discovery protocol in OIDC flow.

Test adds a number of helpers which includes the ability to generate some valid JWK
using the 'github.com/lestrrat-go/jwx/jwk' package.

Introduces a mock test server which simulates a remote keycloak and respinds with
the certs (JWK) and a valid discovery response.