adapter: oidc - Add CreateProvider function for OIDC discovery
This change includes the addition of a constructor for type that will handle
the flow in an OIDC integration with 3scale.

Using the 'github.com/coreos/go-oidc', we accept a URL of an issuer
which allows us to create a Provider for future verification of JWT.

Since we expect the URL to come from 3scale system config, a helper function was
created to strip the basic auth credentials in situatons where they are stored in this format,
since if passed to the Provider, will fail basic verification at the first stage.