adapter: oidc - Provide functionality to handle OIDC integration flow
The OIDC handler will be responsible for taking the header value, along with an issuer URL and
being capable of doing a full verification and parsing of the JWT, extracting what is required by
the handler which is the 'client_id'

At this point, we are defaulting to 'azp' label when the claims are parsed if none are provided.
This is currently bound to Keycloak but will be easily adapted to alternative providers.