authorizer: Introduce package to manage end-to-end authorization
This change introduces a new package with constructs that will
be capble of carrying out full authorization capabilities between
an arbitary service and 3scale.

As of this commit, the 'Manager' struct will be capable
of building out the required clients for 3scale, moving and
reusing the Builder interface that was created previously.

The 'Manager' now pulls in the previously abstracted system
caching functionality which can be invoked or otherwise but
for the caller, it hides the cache and simplifies the interface with
'Manager'