backend: Introduce types for backend caching
This package will house the code required to implement a cacheable
backend solution, external to 3scale itself.

This change introduces some structs:

"Backend" will be an implementation of 3scale client, operating on a
provided cache which will be opaque to the caller.

The "Application" is a local
representation of a model stored in 3scale. It stores counters which can
be used to track rate limiting information, allowing us to make an
authorization decision locally in many cases.