threescale: Update integration and unit tests to support refactor
This change moves away from the previous testing format where we
duplicated the fake http requests in our testing packages and makes use
of the interfaces we created and uses mocks.

This improves the test coverage, but more importantly, cleans up the
tests and makes them more valuable.

This moves out as much testing as possible from the unit tests to the
istio adapter integration testing framework which is more thorough and
tests against a really instance of mixer.

Seperating out the handler allos us to build integration tests with a
variety of inputs as opposed to that hardcoded in the testdata
directory.