backend: Add authorize API to cached backend
This changes satisfies the Authorize API for 3scale client.

Authorizations are made against the linked cache if the key is
available. If we fail to lookup the key and have a cache miss then we
attempt to call out to the remote 3scale to get a known state and
populate the cache with the application before making a grant/deny on
the authoriztion request.