backend: Add AuthRep API to cached backend
This change satisfies the AuthRep API of 3scale client interface.

The behviour is similar to that described previously in the Authorize
API, however in the case of the AuthRep endpoint, where the
authorization has passed and the request has been granted, we take a
write lock and update the counters in the linked cache with the metrics
we were passed in the transaction.