backend: Add Report API to cached backend
This change satisfies the Report API for 3scale client and in doing so
completes the required implementation making the Backend, as a whole a
3scale client proper.

The Report API behaves, from a user perspective, the same as they would
expect when calling a remote 3scale. When calling Report in 3scale,
transactions are added to the queue and processed later, meaning we
would always get an "accepted" response and would only get an error for
a network failure. Since there is no network I/O in this case, a cached
backend should always return "accepted".

Reporting is done against the linked cache where the key is already
known. In cases where it is not known, we act like Auth/AuthRep and do a
request across the network to 3scale to get a known state. However we
may have many applications under a service and if we already know the
hierarchy from another service we can avoid this remote call entirely.

As of now, we don't anticipate much use of this API and so the behaviour
outlined above may change internally later. In particular, we may need
to know a state for each and evey application in full, not just the
hierarchy.