backend: Add Flush functionality to cached backend
The Flush API reports local state to the remote 3scale.
After reporting, it attempts to get the updated state from 3scale for
each application and modify the local state accordingly, depending on
the outcome of both the report, and get updated state request over the
network.

It is expected that Flush will be run at a regular interval as a
background process, however this is left to the caller and will not be
included as part of this package.

The Flush API takes a write lock so may temporarily block incoming
requests to Auth,AuthRep and Report for the same application while doing so.