test:backend: Simulate concurrent requests to Flush tests
This introduces some code to allow the simulation of external requests
during the flushing process.

We need to be able to simulate incoming requests to ensure no
intermediate data is lost during the process and that after Flush has
completed, the cache is in a valid state for the next remote report.