backend: Calculate and store changes to existing limited metrics
This changes keeps track of any new limits that have been added to
already limited metrics.

While handling the sync of state between an existing LimitCounter and a
new LimitCounter as fetched from 3scale, we need to be aware of newly
added limits when deciding to react to an elapsed period as these should
be ignored.