k3s: Stop parsing dependencies for istio version
With this change, we no longer parse the istio version from the
dependencies file. This is due to the fact that updating the
dependencies to build the adapter against the required upstream is
either not worth the effort or will simply not build due to the change upstream
to go modules. Expect istio version to be set in environment or try to
install the latest released version.