authorizer: Support reporting metrics around calls to 3scale
This change adds the ability to gets some data out of the authorizer
with respect to its interactions with 3scale over the network.

Provice a callback function that can be used to hook into the underlying
HTTP response.

This changes all removes the HTTP client builder from the constructor
since it is just needed for testing and we do not expect any other
implementations of this to be passed in from outside the package.